+++ Network/Telegram/API/Bot/Property/Persistable.hs
@@ -1,4 +1,4 @@
-module Network.Telegram.API.Bot.Property.Persistable (Persistable (..), Payload, Capacity (..)) where
+module Network.Telegram.API.Bot.Property.Persistable (Persistable (..), Payload, PL (..), Capacity (..)) where
 
 import "aeson" Data.Aeson (FromJSON, Value, decode, object, (.=))
 import "base" Control.Exception (try)
@@ -6,7 +6,7 @@ import "base" Control.Monad (Monad ((>>=)), join)
 import "base" Data.Function (flip, (.), ($))
 import "base" Data.Functor (Functor (fmap), (<$>))
 import "base" Data.Int (Int, Int64)
-import "base" Data.Maybe (Maybe (Just, Nothing), fromJust)
+import "base" Data.Maybe (fromJust)
 import "base" Data.Semigroup (Semigroup ((<>)))
 import "base" Data.String (String)
 import "base" Data.Tuple (snd)
@@ -18,11 +18,20 @@ import "transformers" Control.Monad.Trans.Reader (ask)
 import "wreq" Network.Wreq.Session (post)
 
 import Network.Telegram.API.Bot.Core (Telegram, Token (Token), Ok, result)
-import Network.Telegram.API.Bot.Object (Keyboard, Notification, Message, Object)
+import Network.Telegram.API.Bot.Object (Object, Keyboard, Notification, Member, Message, Sender)
 
-data Capacity = Post | Edit | Purge
+data Capacity = Fetch | Post | Edit | Purge
 
-type family Payload (c :: Capacity) o = r | r -> c o
+newtype PL (c :: Capacity) (o :: *) a = PL a
+
+type family Payload (c :: Capacity) o = r | r -> o c
+
+type instance Payload 'Post Keyboard = PL 'Post Keyboard (Int64, Int, Keyboard)
+type instance Payload 'Edit Keyboard = PL 'Edit Keyboard (Int64, Int, Keyboard)
+type instance Payload 'Fetch Member = PL 'Fetch Member (Int64, Int)
+type instance Payload 'Purge Message = PL 'Purge Message (Int64, Int)
+type instance Payload 'Post Notification = PL 'Post Notification (Text, Text)
+type instance Payload 'Fetch Sender = PL 'Fetch Sender ()
 
 class Object o => Persistable c o where
 	{-# MINIMAL payload, endpoint #-}
@@ -36,25 +45,28 @@ class Object o => Persistable c o where
 			. fmap (fromJust . join . fmap result . decode @(Ok a) . responseBody)
 				. flip (post session) p $ "https://api.telegram.org/" <> unpack token <> "/" <> e
 
-type instance Payload 'Edit Keyboard = (Int64, Int, Keyboard)
-type instance Payload 'Post Message = (Int64, Text, Maybe Keyboard)
-type instance Payload 'Purge Message = (Int64, Int)
-type instance Payload 'Post Notification = (Text, Text)
-
 instance Persistable 'Edit Keyboard where
-	payload (chat_id, message_id, reply_markup) = object
+	payload (PL (chat_id, message_id, reply_markup)) = object
 		["chat_id" .= chat_id, "message_id" .= message_id, "reply_markup" .= reply_markup]
 	endpoint _ = "editMessageReplyMarkup"
 
-instance Persistable 'Post Message where
-	payload (chat_id, text, Nothing) = object ["chat_id" .= chat_id, "text" .= text]
-	payload (chat_id, text, Just kb) = object ["chat_id" .= chat_id, "text" .= text, "reply_markup" .= kb]
+instance Persistable 'Post Keyboard where
+	payload (PL (chat_id, text, kb)) = object
+		["chat_id" .= chat_id, "text" .= text, "reply_markup" .= kb]
 	endpoint _ = "sendMessage"
 
+instance Persistable 'Fetch Member where
+	payload (PL (chat_id, user_id)) = object ["chat_id" .= chat_id, "user_id" .= user_id]
+	endpoint _ = "getChatMember"
+
 instance Persistable 'Purge Message where
-	payload (chat_id, message_id) = object ["chat_id" .= chat_id, "message_id" .= message_id]
+	payload (PL (chat_id, message_id)) = object ["chat_id" .= chat_id, "message_id" .= message_id]
 	endpoint _ = "deleteMessage"
 
 instance Persistable 'Post Notification where
-	payload (cbq_id, text) = object ["callback_query_id" .= cbq_id, "text" .= text]
+	payload (PL (cbq_id, text)) = object ["callback_query_id" .= cbq_id, "text" .= text]
 	endpoint _ = "answerCallbackQuery"
+
+instance Persistable 'Fetch Sender where
+	payload (PL ()) = object []
+	endpoint _ = "getMe"